backend/requirements.txt
# Backend — Auto Workshop System
# Python 3.11+

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0

sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

python-multipart>=0.0.6
email-validator>=2.0
